refactor(constants): constant index 추가
