+++ apis/index.ts
@@ -1,7 +1,6 @@
 import axios from "axios";
-import { exception } from "@/constants/exception.constant";
+import { ERROR, TOKEN } from "@/constants";
 import { Storage } from "@/storage";
-import { TOKEN } from "@/constants/token.constant";
 import { refreshToken } from "./header";
 
 export const http = axios.create({
@@ -14,7 +13,7 @@ http.interceptors.response.use(
   async (error) => {
     const request = error.config;
     const { code } = error.response.data;
-    const isAccessTokenExpiredError = code === exception.code.TOKEN_403_2;
+    const isAccessTokenExpiredError = code === ERROR.TOKEN_403_2;
 
     if (isAccessTokenExpiredError && !request.sent) {
       request.sent = true;
