+++ constants/error.constant.ts
@@ -1,0 +1,15 @@
+const ERROR = {
+  IMG_400_1: "IMG-400-1",
+  DOCS_404_1: "DOCS-404-1",
+  DOCS_404_2: "DOCS-404-2",
+  DOCS_403_1: "DOCS-403-1",
+  COMMON_403_1: "COMMON-403-1",
+  USER_403_1: "USER-403-1",
+  DOCS_403_2: "DOCS-403-2",
+  TOKEN_403_1: "TOKEN-403-1",
+  TOKEN_403_2: "TOKEN-403-2",
+  TOKEN_403_3: "TOKEN-403-3",
+  USER_404_1: "USER-404-1",
+};
+
+export default ERROR;
