+++ constants/token.constant.ts
@@ -1,4 +1,6 @@
-export const TOKEN = {
+const TOKEN = {
   ACCESS: "access_token",
   REFRESH: "refresh_token",
 } as const;
+
+export default TOKEN;
