+++ storage/index.ts
@@ -1,4 +1,4 @@
-type LocalStorageKey = "access_token" | "refresh_token" | "autoComplete";
+type LocalStorageKey = "access_token" | "refresh_token";
 
 export class Storage {
   private static isWindowAvailable() {
@@ -6,16 +6,16 @@ export class Storage {
   }
 
   static getItem(key: LocalStorageKey) {
-    if (typeof window !== "undefined") return localStorage.getItem(key);
+    if (this.isWindowAvailable()) return localStorage.getItem(key);
   }
 
   static setItem(key: LocalStorageKey, value: string) {
-    if (typeof window === "undefined") return;
+    if (!this.isWindowAvailable()) return;
     localStorage.setItem(key, value);
   }
 
   static delItem(key: LocalStorageKey) {
-    if (typeof window === "undefined") return;
+    if (!this.isWindowAvailable) return;
     localStorage.removeItem(key);
   }
